successful migration of routes from server.js
