+++ backend/routes/index.js
@@ -1,3 +1,8 @@
 const router = require('express').Router()
+const playerRoutes = require('./playerRoutes');
+const teamRoutes = require('./teamRoutes');
 
-router.use('/')
+router.use('/players', playerRoutes);
+router.use('/teams', teamRoutes);
+
+module.exports = router;
